moved home and register to routes dir
